completed certifications

+++ app/routes/catalog.py
@@ -17,9 +17,15 @@ router = APIRouter(
 )
 
 
+def verify_owner(cid, uid, db):
+    
+    biz = db.query(models.Business).filter(models.Business.owner_id == uid).first()
+    cert = db.query(models.Catalog).filter(models.Catalog.id == cid).first()
+    if biz.id == cert.business_id:
+        return True
 
 # ***************ADD CATALOG*******************
-@router.post("/catalog", status_code=status.HTTP_200_OK)
+@router.post("/catalog", status_code=status.HTTP_200_OK, response_model=schemas.CatalogResponse)
 def add_catalog(catalog: schemas.Catalog, db: Session = Depends(get_db), current_user: str = Depends(oauth2.get_current_user)):
 
     query = db.query(models.Business).filter(models.Business.owner_id == current_user.id)
@@ -41,6 +47,9 @@ def update_catalog(id: int, catalog: schemas.Catalog, db: Session = Depends(get_
     query = db.query(models.Catalog).filter(models.Catalog.id == id)
     if query.first() is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Catalog not found")
+    
+    if not verify_owner(id, current_user.id, db):
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail=f"You are not authorized to update this catalog")
     
     query.update(catalog.model_dump(), synchronize_session=False)
     db.commit()
@@ -81,5 +90,20 @@ def delete_catalog(id: int, db: Session = Depends(get_db), current_user: str = D
     if not query.first():
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Catalog not found")
 
+
+    if not verify_owner(id, current_user.id, db):
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail=f"You are not authorized to delete this catalog")
+
+
     query.delete(synchronize_session=False)
     db.commit()
+    return{"data":"deleted"}
+
+
+# ***************GET CATALOG*******************
+@router.get('/catalog/{business_id}', status_code=status.HTTP_200_OK, response_model=List[schemas.CatalogResponse])
+def get_certificate(business_id: int, db: Session = Depends(get_db)):
+    query = db.query(models.Catalog).filter(models.Catalog.business_id == business_id).all()
+    if not query:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"No certificates found")   
+    return query
